+++ packages/client/src/components/live/PreSeat.tsx
